requirements.txt
snowflake-connector-python
snowflake-sqlalchemy
python-dotenv
pandas
polars
pyspark
pyarrow
pytest
pre-commit